fix: change application folder structure
According to conventions "pkg" should stand for modules that should be using exteranly, while "internal" packages - for internal usage. But most of the modules in source code does not make sence outside of the application. That mean that all modules in application ( currently ) only for internal usage and should be moved into "internal" folder.